chore: make a clean clone reproducible, and add the project basics
The test dependencies were declared nowhere. pytest, anyio and ruff had been
installed by hand into the local venv, so `uv sync --locked` on a fresh clone
removed them and left no way to run the suite at all. They are a dependency
group now.

uv.lock did not exist and was gitignored, so every clone resolved whatever was
current that day while the C++ half stayed pinned by CMake -- the two halves of
the build had completely different reproducibility guarantees. It is committed,
and CI runs `uv lock --check` so a pyproject change that was never locked fails
here instead of on someone else's machine.

Also: .python-version, a LICENSE (MIT -- say if you want a different one, it is
your repo), ruff configured and the tree made clean, py.typed, and a Windows CI
workflow.

The suite is split with a `desktop` marker. Those eight tests need a real
display session and a system-wide keyboard hook; a runner has neither in any
meaningful sense, so CI skips them rather than asserting something that does not
mean anything. They still run locally, which is where they matter.

Two of the lint findings were worth having: pytest.raises(Exception) in the
image tests would have passed on a misspelled call, and two match= patterns had
unescaped metacharacters.

+++ tests/conftest.py
@@ -68,7 +68,7 @@ class FakeScreen:
         the aim_ratio a correct calibration should recover.
         """
         dst_w, _ = _native.plan_fit(self.width, self.height, max_w, max_h, False)
-        offset = int(round(self._panned / self.pan_ratio))
+        offset = round(self._panned / self.pan_ratio)
         # A repeating-but-not-periodic pattern, so a shift is unambiguous.
         return [((x + offset) * 37) % 251 + ((x + offset) * 7) % 13 for x in range(dst_w)]
